% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{db_dataset_exists}
\alias{db_dataset_exists}
\title{Check if a hive dataset exists}
\usage{
db_dataset_exists(section, dataset)
}
\arguments{
\item{section}{The section name}

\item{dataset}{The dataset name}
}
\value{
Logical TRUE if exists, FALSE otherwise
}
\description{
Check if a hive dataset exists
}
\examples{
\dontrun{
db_connect()
db_dataset_exists("Trade", "Imports")
# [1] TRUE
}
}
