% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from the CSO Data Lake}
\usage{
db_disconnect()
}
\value{
Invisibly returns TRUE if disconnected, FALSE if was not connected.
}
\description{
Disconnect from the CSO Data Lake
}
\examples{
\dontrun{
db_connect()
# ... do work ...
db_disconnect()
}
}
