% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{db_hive_read}
\alias{db_hive_read}
\title{Read a CSO Dataset from hive-partitioned parquet (lazy)}
\usage{
db_hive_read(section, dataset, ...)
}
\arguments{
\item{section}{The section name (e.g. "Trade")}

\item{dataset}{The dataset name (e.g. "Imports")}

\item{...}{Additional named options passed into DuckDB read_parquet() in SQL form
(e.g. union_by_name = TRUE). See examples below.}
}
\value{
A lazy tbl_duckdb object
}
\description{
Read a CSO Dataset from hive-partitioned parquet (lazy)
}
\examples{
\dontrun{
# Basic read
db_hive_read("Trade", "Imports")

# With options
db_hive_read("Trade", "Imports", union_by_name = TRUE, filename = TRUE)
}
}
