% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{db_list_sections}
\alias{db_list_sections}
\title{List sections in the hive data lake}
\usage{
db_list_sections()
}
\value{
Character vector of section names
}
\description{
Returns all top-level section folders in the data lake.
}
\examples{
\dontrun{
db_connect()
db_list_sections()
# [1] "Trade" "Labour" "Health" "Agriculture"
}
}
