% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{db_list_tables}
\alias{db_list_tables}
\title{List tables in a DuckLake schema}
\usage{
db_list_tables(schema = "main")
}
\arguments{
\item{schema}{Schema name (default "main")}
}
\value{
Character vector of table names
}
\description{
Returns all tables in a given schema.
}
\examples{
\dontrun{
db_lake_connect()
db_list_tables()
# [1] "imports" "exports" "products"

db_list_tables("trade")
# [1] "monthly_summary" "annual_totals"
}
}
