% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{db_list_views}
\alias{db_list_views}
\title{List views in a DuckLake schema}
\usage{
db_list_views(schema = "main")
}
\arguments{
\item{schema}{Schema name (default "main")}
}
\value{
Character vector of view names
}
\description{
Returns all views in a given schema.
}
