% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{db_table_cols}
\alias{db_table_cols}
\title{Get column names for a DuckLake table}
\usage{
db_table_cols(schema = "main", table)
}
\arguments{
\item{schema}{Schema name (default "main")}

\item{table}{Table name}
}
\value{
Character vector of column names
}
\description{
Get column names for a DuckLake table
}
