% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{db_view_cols}
\alias{db_view_cols}
\title{Get column names for a DuckLake view}
\usage{
db_view_cols(schema = "main", view)
}
\arguments{
\item{schema}{Schema name (default "main")}

\item{view}{View name}
}
\value{
Character vector of column names
}
\description{
Get column names for a DuckLake view
}
